requirements.txt
flask==2.0.3
pillow==9.5.0
python-dotenv==0.21.1
requests==2.31.0
werkzeug==2.0.3 